// packages/ide/src/webviews/elf-explorer/src/common/types/metadata.ts
/**
 *
 * Copyright (c) 2024 Analog Devices, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */
export type THeaderInfo = {
	label: string;
	value: string | number;
};

export type TArmAttributes = {
	label: string;
	value: string | number | undefined;
};

export type TMetadata = {
	header: THeaderInfo[];
	armAttributes: TArmAttributes[];
	heuristicInfo: Array<Record<string, any>>;
};